docs: 테스트 목록 수정
