+++ README.md
@@ -34,18 +34,21 @@
 - [x] 우승자 선별
 - [x] 다수의 우승자일 경우, 쉼표로 구분해서 출력. 단일 우승자의 경우, 이름만 출력
 
-# 테스트
+# 테스트 목록
 
-1. Car 객체 테스트
+1. Car
 
-- [x] 객체 생성 테스트
-- [x] 자동차 위치 history 저장 테스트
+- 경주할 자동차 이름 입력 검증
+  - [ ] 자동차는 이름을 가져야한다.
+  - [ ] 자동차 이름은 1~5자여야 한다.
+  - [ ] 자동차는 쉼표를 기준으로 구분한다.
+  - [ ] 자동차는 2대 이상이어야 한다.
+  - [ ] 자동차 이름은 중복되면 안된다.
+- 자동차가 잘 움직이는지 검증
+  - [ ] 자동차는 4 이상의 숫자를 받으면 전진할 수 있다.
 
-2. 검증 테스트
+2. Race
 
-- [x] 자동차 이름 입력 검증 테스트
-- [x] 시도 횟수 입력 검증 테스트
-
-3. 우승자 선별 테스트
-
-- [x] 우승자 선별 테스트
+- 경주 시도 횟수 입력 검증
+  - [ ] 시도 횟수는 1~20 사이여야 한다.
+  - [ ] 시도 횟수는 자연수로 입력해야 한다.
